split IZooKeeperClientExtension
