+++ Vostok.ZooKeeper.Recipes.Tests/DistributedLockToken_Tests.cs
@@ -6,6 +6,7 @@ using Vostok.Commons.Testing;
 using Vostok.ZooKeeper.Client.Abstractions;
 using Vostok.ZooKeeper.Client.Abstractions.Model;
 using Vostok.ZooKeeper.Client.Abstractions.Model.Request;
+using Vostok.ZooKeeper.Recipes.Helpers;
 
 namespace Vostok.ZooKeeper.Recipes.Tests
 {
